Creating sendEmail function on utils
